feat: Add FontAwesome icons and Vue FontAwesome library
This commit adds the FontAwesome icons and the Vue FontAwesome library to the project dependencies. These additions enable the usage of FontAwesome icons within the Vue components, enhancing the visual appeal and improving the overall user experience. The FontAwesome library provides a wide range of high-quality icons in both solid and brand styles. This change also enables the project to leverage the power of FontAwesome's SVG icons throughout the application. The inclusion of FontAwesome and Vue FontAwesome enhances the project's visual aesthetics, making it easier to create engaging and visually appealing UI components.